requirements.txt
amqp==5.3.1
billiard==4.2.4
blinker==1.9.0
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.8.0
Flask==3.1.3
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.2
MarkupSafe==3.0.3
packaging==26.0
prompt_toolkit==3.0.52
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.2.0
six==1.17.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
Werkzeug==3.1.6
